refactored to make the page easier to read
