+++ src/components/applyances/ApplyanceContainer.js
@@ -2,6 +2,9 @@
 import { useState } from "react"
 import { ApplyanceSearch } from "./ApplyanceSearch"
 import { ApplyanceList } from "./ApplyanceList"
+import "./Applyances.css"
+
+
 
 export const ApplyanceContainer = () => {
     const [searchTerms, setSearchTerms] = useState("")
