+++ src/components/applyances/ApplyanceDetails.js
@@ -8,9 +8,7 @@ export const ApplyanceDetails = () => {
     const [applyances, setApplyances] = useState([])
 
 
-    const [applyance, updatedApplyance] = useState({
-
-    })
+    const [applyance, updatedApplyance] = useState({})
 
     const getAllApplyances = () => {
         fetch(`http://localhost:8088/applyances?_sort=makeModel&_expand=user`)
@@ -19,6 +17,7 @@ export const ApplyanceDetails = () => {
                 setApplyances(applyanceArray)
             })
     }
+
     // get apply user object out of local storage
     const localApplyUser = localStorage.getItem("apply_user") // a string
     const applyUserObject = JSON.parse(localApplyUser) // an object with 2 keys (id and staff)
@@ -41,9 +40,9 @@ export const ApplyanceDetails = () => {
     const editButton = () => {
         if (applyUserObject.id === applyance.userId || applyUserObject.admin === true) {
             return <>
-            <footer><Link to={`/all-applyances-edit/${applyanceId}`}><button>
-    Edit ApplYance
-    </button></Link></footer> 
+                <footer><Link to={`/all-applyances-edit/${applyanceId}`}><button>
+                    Edit ApplYance
+                </button></Link></footer>
             </>
         } else {
             return ""
@@ -53,19 +52,19 @@ export const ApplyanceDetails = () => {
     const deleteButton = () => {
         if (applyUserObject.id === applyance.userId || applyUserObject.admin === true) {
             return <>
-            <footer><button
-                        onClick={() => {
-                            fetch(`http://localhost:8088/applyances/${applyanceId}`, {
-                                method: "DELETE"
-                                }) 
+                <footer><button
+                    onClick={() => {
+                        fetch(`http://localhost:8088/applyances/${applyanceId}`, {
+                            method: "DELETE"
+                        })
                             .then(() => {
-                            getAllApplyances()
+                                getAllApplyances()
                             })
-                        }}
-            
-            >
-    Delete ApplYance
-    </button></footer> 
+                    }}
+
+                >
+                    Delete ApplYance
+                </button></footer>
             </>
         } else {
             return ""
@@ -73,24 +72,36 @@ export const ApplyanceDetails = () => {
     }
 
 
-    return (<article className="applyanceArticle">
-    <section className="applyanceSection">
-        <header className="applyance__header">{applyance.makeModel}</header>
-        <div>
-            <img src={`${applyance.picture}`}></img>
-        </div>
-        <div className="manualLink">
-            <Link to={applyance.manual}>Link to Manual</Link>
-        </div>
-        <div>Owner: {applyance?.user?.firstName}</div>
-        <div className="purchaseDiv">Purchase Information:
-            <div>Purchase Date: {applyance.purchaseDate}</div>
-            <div>Purchase Price: {applyance.purchasePrice}</div>
-            <div>Purchase Location: {applyance.purchaseLocation}</div>
-            <div className="tag">{applyance?.tag?.location}</div>
-        </div>
-        {editButton()}
-        {deleteButton()}
-    </section>
+
+    return (<article className="detailsArticle">
+        <section className="applyanceSection">
+
+            <div>
+                <header className="applyance__header">{applyance.makeModel}</header>
+                <div className="tag">{applyance?.tag?.location}</div>
+                <div className="manualLink">
+                    <Link to={applyance.manual}>Link to Manual</Link>
+                </div>
+            </div>
+
+            <div>
+                <img src={`${applyance.picture}`}></img>
+            </div>
+
+            <div>
+                <div>Owner: {applyance?.user?.firstName}</div>
+                <div className="purchaseDiv">Purchase Information:
+                    <div>Purchase Date: {applyance.purchaseDate}</div>
+                    <div>Purchase Price: {applyance.purchasePrice}</div>
+                    <div>Purchase Location: {applyance.purchaseLocation}</div>
+                </div>
+            </div>
+
+            <footer>
+                {editButton()}
+                {deleteButton()}
+            </footer>
+        </section>
+        
     </article>)
 }
