+++ src/components/applyances/ApplyanceForm.js
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom"
+import "./Applyances.css"
+
 
 export const ApplyanceForm = () => {
