+++ src/components/applyances/Applyances.css
@@ -19,10 +19,23 @@ margin-bottom: auto;
     border: 2px solid black;
     padding: 10px;
     display: flex;
-    max-width: 500px;
+    max-width: 300px;
     min-height: 300px;
     flex-direction: column;
-    justify-content: space-between;
+    justify-content: space-between
     
 }
 
+.tag {
+    background: #e63946;
+    color: #f1faee;
+    border-radius: 5px;
+    text-align: center;
+    width: 125px;
+}
+
+footer {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
