+++ src/components/myApplyances/MyApplyanceContainer.js
@@ -2,6 +2,8 @@
 import { useState } from "react"
 import { MyApplyanceList } from "./MyApplyanceList"
 import { MyApplyanceSearch } from "./MyApplyanceSearch"
+import "../applyances/Applyances.css"
+
 
 export const MyApplyanceContainer = () => {
     const [mySearchTerms, mySetSearchTerms] = useState("")
